requirements.txt
torch
torchvision
faiss-cpu
streamlit
Pillow
numpy
pandas
huggingface_hub